Make the URLs change when selecting messages
To do this, I had to come up with some hacky way to not reload the the
page when the url is change. I found it online somewhere and appended it
to the app code - I still have to review it because I'm not particularly
sure what kind of side effects it might have (especially when
transitioning between two unrelated pages, like /login and /messages. I
guess only time will tell, for now.