Implement route authorization
I renamed the user service to auth because that's a lot more descriptive
in terms of what it does. I also integrated it in the routechangestart
event, so that it checks if users are logged in before allowing them to go
about their business. In the auth servie, I implemented an (incomplete)
session-based storage for the session Id - since we have no server, it
really doesn't expire, and loggedIn-ness is determined if WR-session-id is
there or the auth user is there.