Link reply function from controller scope to directive.
I'm not sure what hte best practice is to pass a function down like that,
but I think it should be fine for now. I might have to update some
phrasing in the convo cards (mot likely will change footer to controls or
something) but it seems like an okay tradeoff, and at least it's
accountable.