Forward Card now only appears when the forward button is pressed
I did some stuff with ng-show to make the forward card stay hidden by
default, only appearing on alternate presses of the forward button. This
implementation is currently incomplete, however, because I still want
the recipients form to take focus consistently and for the page to
scroll to the bottom when the forward card appears.