Add functions that assist in user search
I refactored a good part of the userService code. Here are my changes:

1. I benched the fuzzyFindByUsername until I can double check the
levenshtein algorithm - replaced it in controllers with findByUsername
2. I implement findByUsername and findByEmail that tests for exact match
of those fields
3. I implemented a queryByUsername that tests for a (like - startsWith)
match of the username field
4. I restructured everything into a promise format and made sure the
searches had a consistent return format - an object with a score, username
and id.

I made change [1] because the application was slowing down a lot,
calculating the edit distance. startsWith is almost as functionally good,
so I went with that.
[4] was a good thing because everything wasn't really agreeing with each
other, and they shouuuld be drop-ins for each other.