Complete GUI Design of Options Modal
-GUI CSS design complete (include window responsive functionality)
-Update TODO

+What's left is the functionality which would allow the user to change
the data
+Optional feature: Add 'reset' state to modal; when closed, the modal
saves its last 'state' and hence is does not reset to an initial 'state'
when the website is not refreshed

Also, just incase, I took the icons from the internet because I don't
have a good photo editor on my laptop. Not sure if using those icons in
the completed website would cause copyrighted issues. So I'm just
letting you guys know.